hello.go
package main

import "fmt"

const spanish = "Spanish"
const french = "French"
const indonesia = "Indonesia"
const englishHelloPrefix = "Hello, "
const spanishHelloPrefix = "Hola, "
const frenchHelloPrefix = "Bonjour, "
const indonesiaHelloPrefix = "Halo, "

// Hello func only receive string and return string
func Hello(name string, language string) string {
	if name == "" {
		name = "world"
	}

	return greetingPrefix(language) + name
}

func greetingPrefix(language string) (prefix string) {

	switch language {
	case french:
		prefix = frenchHelloPrefix
	case spanish:
		prefix = spanishHelloPrefix
	case indonesia:
		prefix = indonesiaHelloPrefix
	default:
		prefix = englishHelloPrefix
	}

	return
}

func main() {
	fmt.Println(Hello("Ardi", ""))
}
